requirements.txt
openai-whisper==20231117
pydub==0.25.1
python-dotenv==1.0.0
Levenshtein==0.23.0